# public/docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "GadgetByteNepal",
        "description": "GadgetByte Nepal Rest Apis Documentation",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/v1/posts": {
            "get": {
                "tags": [
                    "Posts"
                ],
                "summary": "Get a list of Posts",
                "operationId": "7b253f6aade6833019f938c1f2ef08ed",
                "parameters": [
                    {
                        "name": "q",
                        "in": "query",
                        "description": "Search term for filtering by name ",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "per_page",
                        "in": "query",
                        "description": "Number of items per page (optional, default: 20)",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "default": 20
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/PostResource"
                                    }
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request"
                    }
                },
                "security": [
                    {
                        "bearer_token": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Posts"
                ],
                "summary": "Create a new Post",
                "operationId": "b996500c6caf55100cd024d3e3b49dd5",
                "parameters": [
                    {
                        "name": "X-User-Id",
                        "in": "header",
                        "description": "User ID for authentication",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "requestBody": {
                    "description": "Posts data",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/StorePostRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Successfully created post",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/PostResource"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Unprocessable Entity (Validation error)",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ValidationErrorResponse"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request"
                    }
                },
                "security": [
                    {
                        "bearer_token": []
                    },
                    {
                        "X-User-Id": []
                    }
                ]
            }
        },
        "/api/v1/posts/{id}": {
            "get": {
                "tags": [
                    "Posts"
                ],
                "summary": "Get a specific post",
                "operationId": "db686e4ed73ae7ca2424e0a4f740c8b2",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "The ID of the post to retrieve",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/PostResource"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Not Found"
                    }
                },
                "security": [
                    {
                        "bearer_token": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "Posts"
                ],
                "summary": "Update an existing post",
                "description": "Update the specified resource in storage.",
                "operationId": "98c3c32a86125c864caa21999f234493",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "The ID of the post to update",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "X-User-Id",
                        "in": "header",
                        "description": "User ID for authentication",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "requestBody": {
                    "description": "Post data",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/UpdatePostRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Successfully updated post",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/PostResource"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Unprocessable Entity (Validation error)",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ValidationErrorResponse"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Not Found"
                    }
                },
                "security": [
                    {
                        "bearer_token": []
                    },
                    {
                        "X-User-Id": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Posts"
                ],
                "summary": "Delete an post",
                "description": "Remove the specified resource from storage.",
                "operationId": "82927d2153173c1a7d6ab1e9d52ff56e",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "The ID of the post to delete",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "X-User-Id",
                        "in": "header",
                        "description": "User ID for authentication",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successfully deleted post",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "string",
                                            "example": "success"
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Post deleted successfully"
                                        },
                                        "data": {
                                            "$ref": "#/components/schemas/PostResource"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Not Found"
                    }
                },
                "security": [
                    {
                        "bearer_token": []
                    },
                    {
                        "X-User-Id": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
            "ValidationErrorResponse": {
                "title": "Validation Error Response",
                "description": "An object representing a validation error response",
                "properties": {
                    "message": {
                        "description": "The error message",
                        "type": "string",
                        "example": "The given data was invalid."
                    },
                    "errors": {
                        "description": "An object containing field validation error messages. (This may varies from your api endpoints.)",
                        "type": "object",
                        "example": {
                            "postkey": [
                                "The postkey field is required."
                            ]
                        }
                    }
                },
                "type": "object"
            },
...
            "StorePostRequest": {
                "title": "Post Store Request",
                "description": "Post store request data",
                "required": [
                    "title",
                    "slug",
                    "description",
                    "short_description",
                    "category_id",
                    "sub_category_id",
                    "isprice",
                    "show_in_pricelist"
                ],
                "properties": {
                    "title": {
                        "description": "The Post Title",
                        "type": "string",
                        "example": "Dashboard"
                    },
                    "slug": {
                        "description": "The Slug for the post",
                        "type": "string",
                        "example": "apple"
                    },
                    "short_description": {
                        "description": "The Post's short description ",
                        "type": "string",
                        "example": "this is the short description of it."
                    },
                    "description": {
                        "description": "The Post's description ",
                        "type": "string",
                        "example": "this is the description of it."
                    },
                    "category_id": {
                        "description": "The  Post Category",
                        "type": "integer",
                        "example": "1"
                    },
                    "sub_category_id": {
                        "description": "The Post Sub Category",
                        "type": "integer",
                        "example": "1"
                    },
                    "isprice": {
                        "description": "If Price then 'Y' else 'N'",
                        "type": "string",
                        "example": "Y"
                    },
                    "show_in_pricelist": {
                        "description": "If Price to be shown for price list  then 'Y' else 'N'",
                        "type": "string",
                        "example": "Y"
                    },
                    "tags": {
                        "description": "Ids of the tag",
                        "type": "array",
                        "items": {
                            "type": "integer"
                        }
                    }
                },
                "type": "object"
            },
...
            "UpdatePostRequest": {
                "title": "Post Update Request",
                "description": "Post update request data",
                "required": [
                    "title",
                    "slug",
                    "description",
                    "short_description",
                    "category_id",
                    "sub_category_id",
                    "isprice",
                    "show_in_pricelist"
                ],
                "properties": {
                    "title": {
                        "description": "The Post Title",
                        "type": "string",
                        "example": "Dashboard"
                    },
                    "slug": {
                        "description": "The Slug for the post",
                        "type": "string",
                        "example": "apple"
                    },
                    "short_description": {
                        "description": "The Post's short description ",
                        "type": "string",
                        "example": "this is the short description of it."
                    },
                    "description": {
                        "description": "The Post's description ",
                        "type": "string",
                        "example": "this is the description of it."
                    },
                    "category_id": {
                        "description": "The  Post Category",
                        "type": "integer",
                        "example": "1"
                    },
                    "sub_category_id": {
                        "description": "The Post Sub Category",
                        "type": "integer",
                        "example": "1"
                    },
                    "isprice": {
                        "description": "If Price then 'Y' else 'N'",
                        "type": "string",
                        "example": "Y"
                    },
                    "show_in_pricelist": {
                        "description": "If Price to be shown for price list  then 'Y' else 'N'",
                        "type": "string",
                        "example": "Y"
                    },
                    "tags": {
                        "description": "Ids of the tag",
                        "type": "array",
                        "items": {
                            "type": "integer"
                        }
                    }
                },
                "type": "object"
            },
...
            "PostResource": {
                "title": "Post Resource",
                "description": "Post resource representation",
                "properties": {
                    "id": {
                        "description": "The unique id of the post",
                        "type": "integer",
                        "example": "1"
                    },
                    "title": {
                        "description": "The display title of the post",
                        "type": "string",
                        "example": "Samsung Phone Nepal"
                    },
                    "slug": {
                        "description": "The Slug of the post",
                        "type": "string",
                        "example": "samsung-phones-nepal"
                    },
                    "short_description": {
                        "description": "The Short Description of the post",
                        "type": "string",
                        "example": "short description goes here"
                    },
                    "description": {
                        "description": "The Description of the post",
                        "type": "string",
                        "example": "Description goes here"
                    },
                    "category_id": {
                        "description": "The Category of the post",
                        "type": "integer",
                        "example": "1"
                    },
                    "sub_category_id": {
                        "description": "The Sub Category of the post",
                        "type": "integer",
                        "example": "3"
                    }
                },
                "type": "object"
            },
...
        },
        "securitySchemes": {
            "bearer_token": {
                "type": "http",
                "description": "Bearer Token Authentication",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}